### 2025-03-15 #7883 @danielskatz
Also, your paper is over 2700 words, while JOSS papers are supposed to be no more than around 1000 words.  You thus need to reduce the size of your paper by quite a bit, or alternatively, perhaps JOSS is not the right venue for what you intend to publish.
