+++ paper.md
@@ -1,0 +1,205 @@
+---
+title: 'QuantumReservoirPy: A Software Package for Time Series Prediction'
+tags:
+  - Python
+  - physics
+  - quantum information
+  - auantum circuits
+  - qiskit
+authors:
+  - name: Ola Tangen Kulseng
+    orcid: 0009-0009-9807-4975
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 1
+  - name: Stanley Miao
+    equal-contrib: true
+    affiliation: 2 # (Multiple affiliations must be quoted)
+  - name: Franz G. Fuchs
+    orcid: 0000-0003-3558-503X
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 3
+  - name: Alexander Stasik
+    orcid: 0000-0003-1646-2472
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: 3, 4
+affiliations:
+ - name: David R. Cheriton School of Computer Science, University of Waterloo, Canada
+   index: 2
+ - name: Department of Physics, Norwegian University of Science and Technology (NTNU), Norway
+   index: 1
+ - name: Department of Mathematics and Cybernetics, SINTEF Digital, Norway
+   index: 3
+ - name: Department of Data Science, Norwegian University of Life Science, Norway
+   index: 4
+date: 22 February 2025
+bibliography: paper.bib
+---
+
+# Summary
+Reservoir computing on quantum computers have recently emerged as a potential resource for time series prediction, owing to their inherent complex dynamics. To advance Quantum Reservoir Computing (QRC) research, we have developed the Python package [`QuantumReservoirPy`](https://github.com/OpenQuantumComputing/quantumreservoirpy), which fascilitates QRC using quantum circuits as reservoirs.
+
+The package is designed to be easy to use, while staying completely customizable. The resulting interface, similar to that of [reservoirpy](https://github.com/reservoirpy/reservoirpy) [@trouvain20], simplifies the development of quantum reservoirs, and provides logical methods of comparison between reservoir architectures.
+
+
+<!-- Hence, there is a need for a flexible framework to test quantum circuits as nonlinear dynamical systems. -->
+
+<!-- We have developed a software package to allow for quantum reservoirs to fit a common structure, similar to that of reservoirpy [@trouvain20], which is advertised as
+ "a python tool designed to easily define, train and use [classical] reservoir computing architectures". -->
+
+<!--
+Our package results in simplified development and logical methods of comparison between quantum reservoir architectures.
+
+Examples are provided to demonstrate the resulting simplicity of executing quantum reservoir computing using our software package. -->
+
+# Statement of need
+Reservoir computing (RC) is a paradigm in machine learning for time series prediction. With recent developments, it has shown a promising advantage in efficiency due to the relative simplicity of the associated training process over conventional neural network methods [@tanaka2019].
+In reservoir computing, a dynamical system composed of hidden functional representations with non-linear state transitions is chosen as a reservoir. Input data from a time series is sequentially encoded and fed into the reservoir. The hidden parameters of the reservoir undergo a non-linear evolution dependent on the stored state and the encoded information fed into the system. Features are subsequently decoded from a readout of certain parameters of the reservoir, which is used to train a simple linear model for the desired time series prediction output, see \autoref{fig1} for an illustration.
+
+![A quantum reservoir system consists of a learning task, an en- and de-coder (red) and the dynamic system itself (green). In standard RC the machine learning part is linear regression.\label{fig1}](fig1.pdf)
+
+
[…]
