### 2025-03-17 #7883 @otkulseng
Hi @danielskatz, I have just made a pull request to merge the paper into the main branch, which should contain all code + paper tomorrow.

As for the word-count, I will have a chat with my coauthors to figure out the next step. I will come back to you as quick as I can.
