+++ paper.bib
@@ -1,0 +1,100 @@
+@InProceedings{fernando03,
+author="Fernando, Chrisantha
+and Sojakka, Sampsa",
+title="Pattern Recognition in a Bucket",
+booktitle="Advances in Artificial Life",
+year="2003",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="588--597",
+abstract="This paper demonstrates that the waves produced on the surface of water can be used as the medium for a ``Liquid State Machine'' that pre-processes inputs so allowing a simple perceptron to solve the XOR problem and undertake speech recognition. Interference between waves allows non-linear parallel computation upon simultaneous sensory inputs. Temporal patterns of stimulation are converted to spatial patterns of water waves upon which a linear discrimination can be made. Whereas Wolfgang Maass' Liquid State Machine requires fine tuning of the spiking neural network parameters, water has inherent self-organising properties such as strong local interactions, time-dependent spread of activation to distant areas, inherent stability to a wide variety of inputs, and high complexity. Water achieves this ``for free'', and does so without the time-consuming computation required by realistic neural models. An analogy is made between water molecules and neurons in a recurrent neural network.",
+isbn="978-3-540-39432-7"
+}
+
+@inproceedings{trouvain20,
+  TITLE = {{ReservoirPy: an Efficient and User-Friendly Library to Design Echo State Networks}},
+  AUTHOR = {Trouvain, Nathan and Pedrelli, Luca and Dinh, Thanh Trung and Hinaut, Xavier},
+  URL = {https://inria.hal.science/hal-02595026},
+  BOOKTITLE = {{ICANN 2020 - 29th International Conference on Artificial Neural Networks}},
+  ADDRESS = {Bratislava, Slovakia},
+  YEAR = {2020},
+  MONTH = Sep,
+  KEYWORDS = {Reservoir Computing ; Echo State Networks ; Offline Learning ; Online Learning ; Hyperparameter Optimization ; Parallel Computing ; Sparse Matrix Computation ; Toolbox},
+  PDF = {https://inria.hal.science/hal-02595026v2/file/Trouvain-et-al2020_ReservoirPy_Icann2020_preprint.pdf},
+  HAL_ID = {hal-02595026},
+  HAL_VERSION = {v2},
+}
+
+@article{jaeger04,
+author = {Herbert Jaeger  and Harald Haas },
+title = {Harnessing Nonlinearity: Predicting Chaotic Systems and Saving Energy in Wireless Communication},
+journal = {Science},
+volume = {304},
+number = {5667},
+pages = {78-80},
+year = {2004},
+doi = {10.1126/science.1091277},
+URL = {https://www.science.org/doi/abs/10.1126/science.1091277},
+eprint = {https://www.science.org/doi/pdf/10.1126/science.1091277},
+abstract = {We present a method for learning nonlinear systems, echo state networks (ESNs). ESNs employ artificial recurrent neural networks in a way that has recently been proposed independently as a learning mechanism in biological brains. The learning method is computationally efficient and easy to use. On a benchmark task of predicting a chaotic time series, accuracy is improved by a factor of 2400 over previous techniques. The potential for engineering applications is illustrated by equalizing a communication channel, where the signal error rate is improved by two orders of magnitude.}}
+
+@Article{mujal23,
+author={Mujal, Pere
+and Mart{\'i}nez-Pe{\~{n}}a, Rodrigo
+and Giorgi, Gian Luca
+and Soriano, Miguel C.
+and Zambrini, Roberta},
+title={Time-series quantum reservoir computing with weak and projective measurements},
+journal={npj Quantum Information},
+year={2023},
+month={2},
+day={23},
+volume={9},
+number={1},
+pages={16},
[…]
