### 2025-03-31 4f607a3
shortened paper to 1000 words for JOSS
