+++ paper.md
@@ -42,39 +42,20 @@ Reservoir computing on quantum computers have recently emerged as a potential re
 The package is designed to be easy to use, while staying completely customizable. The resulting interface, similar to that of [reservoirpy](https://github.com/reservoirpy/reservoirpy) [@trouvain20], simplifies the development of quantum reservoirs, and provides logical methods of comparison between reservoir architectures.
 
 
-<!-- Hence, there is a need for a flexible framework to test quantum circuits as nonlinear dynamical systems. -->
-
-<!-- We have developed a software package to allow for quantum reservoirs to fit a common structure, similar to that of reservoirpy [@trouvain20], which is advertised as
- "a python tool designed to easily define, train and use [classical] reservoir computing architectures". -->
-
-<!--
-Our package results in simplified development and logical methods of comparison between quantum reservoir architectures.
-
-Examples are provided to demonstrate the resulting simplicity of executing quantum reservoir computing using our software package. -->
-
 # Statement of need
-Reservoir computing (RC) is a paradigm in machine learning for time series prediction. With recent developments, it has shown a promising advantage in efficiency due to the relative simplicity of the associated training process over conventional neural network methods [@tanaka2019]. At the core of reservoir computing is the idea of a reservoir, which is a dynamical system with non-linear state transitions that is used to embed time series data into a high dimensional space.
+Reservoir computing (RC) is a paradigm in machine learning for time series prediction. With recent developments, it has shown a promising efficacy compared to conventional neural networks, owing to their relatively simple training process [@tanaka2019].
 
-<!-- In reservoir computing, a dynamical system composed of hidden functional representations with non-linear state transitions is used to embed time series data in a high-dimensional space. -->
+The main building block of RC is a dynamical system called a reservoir. By making the dynamics of the reservoir dependent on the input time series, the state of the reservoir would become a high-dimensional, non-linear transformation of the time series. The hope is that such a low-to-high dimensional encoding enables forecasting using a relatively simple method like Ridge regression. A reservoir can be virtual, such as a sparsely-connected recurrent neural network with random fixed weights, termed as an echo state network [@jaeger04] or even physical, such as a bucket of water [@fernando03]. See \autoref{fig1} for an illustration of a typical RC pipeline.
 
- Input data from the time series is sequentially encoded and fed into the reservoir. The hidden parameters of the reservoir undergo a non-linear evolution dependent on the stored state and the encoded information fed into the system. Features are subsequently decoded from a readout of certain parameters of the reservoir, which is used to train a simple linear model for the desired time series prediction output, see \autoref{fig1} for an illustration.
 
 ![A quantum reservoir system consists of a learning task, an en- and de-coder (red) and the dynamic system itself (green). In standard RC the machine learning part is linear regression.\label{fig1}](fig1.pdf)
 
 
-The selected reservoir must incorporate non-linear state transitions within the limitations of a fixed structure. A reservoir can be virtual, such as a sparsely-connected recurrent neural network with random fixed weights, termed as an echo state network [@jaeger04] or even physical, such as a bucket of water [@fernando03].
-
-<!-- RC can be phrased as a supervised machine learning problem. Typically, one has an observed input sequence $\{x_t\}$ and a target sequence $\{y_t\}$. The reservoir $f$ then performs an evolution in time given by $ u_{t+1} = f(u_t, x_t)$
-where $u_t$ is the internal state of the reservoir at time $t$. The output of the reservoir is given by $\hat{y}_t = W_\text{out} h(u_t)$
[…]
