### 2025-03-31 #7883 @otkulseng
> [@otkulseng](https://github.com/otkulseng) - this has passed our scope review. So now, please do cut the paper down to close to 1000 words maximum. Then I will look for an editor to handle the review.

@danielskatz Glad to hear. I have now shortened the paper to approximately 1000 words, disregarding the 'author'-specification on the top.
