### 2025-03-31 #7883 @danielskatz
👋 @kareem1925, @jameshorine, @obliviateandsurrender - Would one or more of you be willing to review this submission for JOSS?
