### 2025-04-01 #7883 @danielskatz
👋 @rudraprsd and @Abinashbunty - Would one or both of you be willing to review this submission for JOSS?
