### 2025-04-02 #7994 @editorialbot
[REVIEW]: QuantumReservoirPy: A Software Package for Time Series Prediction
**Submitting author:** @otkulseng (<a href="http://orcid.org/0009-0009-9807-4975">Ola Tangen Kulseng</a>)
**Repository:** https://github.com/OpenQuantumComputing/quantumreservoirpy
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2
