### 2025-04-12 OpenQuantumComputing/quantumreservoirpy#26 @rudraprsd
[JOSS] Missing Contribution and Support Guidelines
Hi,

I'm currently reviewing this project as part of the JOSS review process [review thread](https://github.com/openjournals/joss-reviews/issues/7994#issuecomment-2773507931).

I noticed that the repository does not include a CONTRIBUTING.md file or any guidance in the README for how to contribute, report issues, or seek support. Including this information would help improve accessibility and community involvement.

Thank you.
