### 2025-05-13 OpenQuantumComputing/quantumreservoirpy#27 @Abinashbunty
Reviewer feedback for JOSS
Hi authors and contributors, as part of the ongoing review process of your repository in https://github.com/openjournals/joss-reviews/issues/7994, I noted down some of the feedback/suggestions below:
- Tag should be `quantum circuits`
https://github.com/OpenQuantumComputing/quantumreservoirpy/blob/e3af8d98327900bf411926ab7709cd36017bf073/paper.md?plain=1#L7

- Spell check for the word `facilitates`.
https://github.com/OpenQuantumComputing/quantumreservoirpy/blob/e3af8d98327900bf411926ab7709cd36017bf073/paper.md?plain=1#L40

- DOI fix for the `ReservoirPy` paper. The below can most likely be used as the reference for the paper so that the JOSS bot warning is gone.

```
@InProceedings{10.1007/978-3-030-61616-8_40,
author="Trouvain, Nathan
and Pedrelli, Luca
and Dinh, Thanh Trung
and Hinaut, Xavier",
editor="Farka{\v{s}}, Igor
and Masulli, Paolo
and Wermter, Stefan",
title="ReservoirPy: An Efficient and User-Friendly Library to Design Echo State Networks",
booktitle="Artificial Neural Networks and Machine Learning -- ICANN 2020",
year="2020",
publisher="Springer International Publishing",
address="Cham",
pages="494--505",
abstract="We present a simple user-friendly library called ReservoirPy based on Python scientific modules. It provides a flexible interface to implement efficient Reservoir Computing (RC) architectures with a particular focus on Echo State Networks (ESN). Advanced features of ReservoirPy allow to improve up to {\$}{\$}87.9{\backslash}{\%}{\$}{\$}of computation time efficiency on a simple laptop compared to basic Python implementation. Overall, we provide tutorials for hyperparameters tuning, offline and online training, fast spectral initialization, parallel and sparse matrix computation on various tasks (MackeyGlass and audio recognition tasks). In particular, we provide graphical tools to easily explore hyperparameters using random search with the help of the hyperopt library.",
isbn="978-3-030-61616-8"
}
```
