### 2025-05-13 OpenQuantumComputing/quantumreservoirpy#27 @Abinashbunty
In addition to the above, there are couple of more suggestions:

- The docs for the functions can be more elaborate for users or contributors of the package. The relevant functions do not have sufficient docstrings, which might not help a new user to use the package efficiently.
- It will be helpful to have the versions of Python the project supports so that the package works straight out of box for any user with the necessary compatible Python versions.
