### 2025-05-13 #7994 @Abinashbunty
@danielskatz I have completed my review. The functionality and installation works fine for me. I have mainly provided feedback about inclusion of more documentation for core functions to help a new user use the package efficiently. Also provided miscellaneous spell check fixes and DOI fix for the JOSS bot warning.

Once these are addressed, I can mark the pending 2 boxes in the checklist. 👍🏻
