### 2025-05-14 #7994 @danielskatz
Yes, this is reasonable.  Good luck finishing your thesis!
