+++ examples/stabilizer/run_tableaus.sh
@@ -2,7 +2,6 @@
 
 for j in {1..100}
 do
-    #bash run_reservoir.sh 5 3 20 quantum_stab None 2 1 logistic $j
-    bash run_reservoir.sh 3 2 20 quantum_stab None 2 1 logistic $j
+    bash run_reservoir.sh 5 3 20 quantum_stab None 2 1 logistic $j
 done
 
