### 2025-05-19 8dba784
Random stab code (#21)
* random stabilizer decoder

* added creation of random tableaus

* run random stabilizers

---------

Co-authored-by: deltark <dmartres@gmail.com>
