+++ quantumreservoirpy/stabilizer.py
@@ -1,23 +1,31 @@
 from itertools import combinations, product
 import numpy as np
-from qiskit import QuantumCircuit, ClassicalRegister, QuantumRegister
+from qiskit import QuantumCircuit, ClassicalRegister, QuantumRegister, AncillaRegister
+from qiskit.quantum_info import random_clifford, Clifford, Pauli
+from qiskit.circuit.library import PauliEvolutionGate
 
 from quantumreservoirpy.util import randomIsing, get_Ising_circuit
 from quantumreservoirpy.reservoirs import Static
+
+from typing import Iterable
 
 
 class Stabilizer(Static):
     def __init__(
         self,
         n_qubits,
-        n_meas,
+        n_meas,  # number of stabilizer generators
+        tableau: dict | None = None,  # if specified, overrides the tableau generation
+        codestate_preparation_circ: (
+            Iterable[QuantumCircuit] | None
+        ) = None,  # if None, will generate a random stabilizer code
         memory=np.inf,
         backend=None,
         degree=1,
         num_reservoirs=1,
         standard=False,
         isingparams=None,
-        decode=True,# danger zone: this is only for testing
+        decode=True,  # danger zone: this is only for testing
     ) -> None:
         super().__init__(
             n_qubits + 1, memory, backend, degree=degree, num_reservoirs=num_reservoirs
@@ -28,7 +36,7 @@ class Stabilizer(Static):
         if not isingparams:
             steps = 1
             dt = 1.645
-            #top = limitrange(list(combinations(range(n_qubits), 2)))
+            # top = limitrange(list(combinations(range(n_qubits), 2)))
             top = list(combinations(range(n_qubits), 2))
             self.U = {}
             self.isingparams = {}
@@ -42,8 +50,23 @@ class Stabilizer(Static):
             for nr in range(1, num_reservoirs + 1):
                 self.U[nr] = get_Ising_circuit(n_qubits, isingparams[nr])
 
-        self.cs = Stabilizer.get_stabilizer_circuits(n_qubits, n_meas, random=False, standard=standard)
-        self.decodermap = Stabilizer.build_decoder_map(n_meas + 1, standard=standard)
+        if tableau != None:
+            if (
+                len(tableau["stabilizer"]) != n_meas
+                or len(tableau["destabilizer"]) != n_meas
+            ):
+                raise Exception(
+                    "Error: the number of entries of the tableau does not match the dimension of the stabilizer"
+                )
+            self.tableau = tableau
+        else:
+            self.tableau = Stabilizer.generate_tableau(
+                n_qubits, n_meas, codestate_preparation_circ
+            )
+
+    def before(self, circuit):
+        if self.decode:
+            circuit.add_register(AncillaRegister(self.n_meas))
 
     def during(self, circuit, timestep, reservoirnumber):
         circuit.barrier()
@@ -51,8 +74,9 @@ class Stabilizer(Static):
         # encode
         for k in range(self.n_meas):
             beta = 3**k
-            circuit.rx(-beta / 2 * np.pi * timestep, k)
-        # circuit.rx(np.pi * timestep, 0)
+            pauliop = Pauli(self.tableau["destabilizer"][k])
+            evo = PauliEvolutionGate(pauliop, -beta / 2 * np.pi * timestep)
+            circuit.append(evo, range(self.n_qubits - 1))
 
         # reservoir
         circuit.append(self.U[reservoirnumber], range(self.n_qubits - 1))
@@ -60,14 +84,61 @@ class Stabilizer(Static):
         # decode
         cr = ClassicalRegister(self.n_meas)
         circuit.add_register(cr)
-        circuit.barrier()
-        for j in range(self.n_meas):
[…]
