### 2025-05-21 #7994 @danielskatz
👋 @akashkthkr - thanks, but we already have 3 reviewers, and the JOSS requirement is two, so we don't really need another. Of course, you can review it and contribute via issues/PRs in the source repo outside of the JOSS process, just as with any other open source software package.
