+++ CONTRIBUTING.md
@@ -1,0 +1,13 @@
+# Contributing to This Project
+
+Thank you for considering contributing! We welcome all kinds of contributions—bug reports, feature requests, documentation, tests, and code.
+
+---
+
+## How to Contribute
+Feel free to submit a "pull request" on GitHub in a designated feature branch. Include comments about what the fix/update does.
+
+For new features, please also consider updating the examples folder to show the intended use of the feature.
+
+## Contact
+For questions about the package, you can find contact info [in the documentation.](https://quantumreservoirpy.readthedocs.io/)
