+++ paper.bib
@@ -11,18 +11,22 @@ abstract="This paper demonstrates that the waves produced on the surface of wate
 isbn="978-3-540-39432-7"
 }
 
-@inproceedings{trouvain20,
-  TITLE = {{ReservoirPy: an Efficient and User-Friendly Library to Design Echo State Networks}},
-  AUTHOR = {Trouvain, Nathan and Pedrelli, Luca and Dinh, Thanh Trung and Hinaut, Xavier},
-  URL = {https://inria.hal.science/hal-02595026},
-  BOOKTITLE = {{ICANN 2020 - 29th International Conference on Artificial Neural Networks}},
-  ADDRESS = {Bratislava, Slovakia},
-  YEAR = {2020},
-  MONTH = Sep,
-  KEYWORDS = {Reservoir Computing ; Echo State Networks ; Offline Learning ; Online Learning ; Hyperparameter Optimization ; Parallel Computing ; Sparse Matrix Computation ; Toolbox},
-  PDF = {https://inria.hal.science/hal-02595026v2/file/Trouvain-et-al2020_ReservoirPy_Icann2020_preprint.pdf},
-  HAL_ID = {hal-02595026},
-  HAL_VERSION = {v2},
+@InProceedings{trouvain20,
+author="Trouvain, Nathan
+and Pedrelli, Luca
+and Dinh, Thanh Trung
+and Hinaut, Xavier",
+editor="Farka{\v{s}}, Igor
+and Masulli, Paolo
+and Wermter, Stefan",
+title="ReservoirPy: An Efficient and User-Friendly Library to Design Echo State Networks",
+booktitle="Artificial Neural Networks and Machine Learning -- ICANN 2020",
+year="2020",
+publisher="Springer International Publishing",
+address="Cham",
+pages="494--505",
+abstract="We present a simple user-friendly library called ReservoirPy based on Python scientific modules. It provides a flexible interface to implement efficient Reservoir Computing (RC) architectures with a particular focus on Echo State Networks (ESN). Advanced features of ReservoirPy allow to improve up to {\$}{\$}87.9{\backslash}{\%}{\$}{\$}of computation time efficiency on a simple laptop compared to basic Python implementation. Overall, we provide tutorials for hyperparameters tuning, offline and online training, fast spectral initialization, parallel and sparse matrix computation on various tasks (MackeyGlass and audio recognition tasks). In particular, we provide graphical tools to easily explore hyperparameters using random search with the help of the hyperopt library.",
+isbn="978-3-030-61616-8"
 }
 
 @article{jaeger04,
