### 2025-06-03 OpenQuantumComputing/quantumreservoirpy#26 @otkulseng
Hi,
thank you for the feedback. PR on the way.
