### 2025-06-03 81bb646
Docstrings and python version
