### 2025-06-03 bc7a0e7
Docstrings and python version
