### 2025-06-04 #7994 @Abinashbunty
@danielskatz The author has implemented the changes that were suggested. So all items are checked off from my side and good to go. 🚀  The editorialbot still provides an error for one particular DOI even though the author has referenced it correctly. Not sure if there is any fix to resolve this.
