### 2025-06-04 #7994 @obliviateandsurrender
Looks good to go from my end as well as soon as the following two are addressed!

1. In the paper where dependencies have been listed, please include the version numbers with which it is currently compatible. For example, Qiskit often deprecates a significant amount of functionality with its subsequent releases, making it easier for a reader to assess.
2. I would encourage you to add a test suite for your work, at least a basic one, to have an automated way of knowing if anything breaks during upgrades.
3. [Optional] It might be helpful to add the `examples` to your documentation as well.
