### 2025-06-06 #7994 @danielskatz
@otkulseng - it looks like there are still a couple of issues from @obliviateandsurrender for you to address, but that otherwise, this is almost done.

Regarding the DOI error, it seems like you need to add
```doi = {10.1007/978-3-030-61616-8_40},```
in the bib entry for `trouvain20`
