### 2025-06-08 7709802
doi fix and version specifications
