### 2025-06-08 45d71e8
tests and doi fix
