+++ tests/test_imports.py
@@ -1,0 +1,7 @@
+def test_import_main():
+    import quantumreservoirpy
+
+def test_import_submodules():
+    from quantumreservoirpy import music
+    from quantumreservoirpy import reservoirs
+    from quantumreservoirpy import plot
