### 2025-06-08 #7994 @otkulseng
> Amazing work [@otkulseng](https://github.com/otkulseng)! Looks good to go from my end as well as soon as the following two are addressed!
>
> 1. In the paper where dependencies have been listed, please include the version numbers with which it is currently compatible. For example, Qiskit often deprecates a significant amount of functionality with its subsequent releases, making it easier for a reader to assess.
> 2. I would encourage you to add a test suite for your work, at least a basic one, to have an automated way of knowing if anything breaks during upgrades.
> 3. [Optional] It might be helpful to add the `examples` to your documentation as well.

@obliviateandsurrender , please see PR

https://github.com/OpenQuantumComputing/quantumreservoirpy/pull/31

I have added import tests, as well as a working running example to ensure the code compiles. I have also updated the paper to include versioning, to the best of my ability.

We will also consider updating the documentation to include the examples, but that will only be started after the other authors are done with their other projects.

Again, thank you all for your patience.
