### 2025-06-09 #7994 @danielskatz
👋 @obliviateandsurrender - when you get a chance, please see if this now seems ready to accept to you
