### 2025-06-10 03e0c86
changes in paper.md
