### 2025-06-10 #7994 @danielskatz
👋 @otkulseng - I've suggested a number of grammar changes in https://github.com/OpenQuantumComputing/quantumreservoirpy/pull/32 - please merge this, or let me know what you disagree with, then we can continue toward acceptance and publication
