### 2025-06-10 #7994 @otkulseng
@danielskatz , The tagged version is v0.2 and can be found here: https://github.com/OpenQuantumComputing/quantumreservoirpy/releases/tag/v0.2

The doi from zenodo (including the entire repository) is: 10.5281/zenodo.15633043

Please let me know if I have done anything wrong.
