### 2025-06-10 #7994 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
